Land the jfxr sounds, close M5 on Windows and start M6
Three pieces of work that were in the working tree together, in files
they share:

- Sound effects designed in jfxr: golib.NewSoundFile also reads the
  .jfxr files jfxr saves, with a Go version of its synthesizer, and
  golib dist adds jfxr's license to every game.
- M5, Shipping, done on Windows. A Windows game that can't load raylib
  or libffi now says why, in a message box when nobody would see the
  console, instead of ending silently; debug builds carry the game's
  icon and details; golib build and shot work while golib run has the
  game open, golib works while a copy of it is running, and clean says
  what it couldn't delete. What is left of shipping is about other
  platforms, now gathered in one section of the roadmap, on hold.
- M6, 2D essentials, started. Vectors and a 2D camera, saving high
  scores and progress, looping sounds, outlines, polygons, aligned
  text, WithOpacity, SetMouseVisible and Input.MouseMoved; golib test
  <game>; a template whose world is not the screen. Two games made by
  agents from a single prompt each, games/skyraid and games/crates,
  showed what was missing and stay as examples.

+++ tools/cli/project.go
@@ -100,7 +100,9 @@ func waitForGame(cmd *exec.Cmd, timeout time.Duration) (exitCode int, timedOut b
 	}
 	var exit *exec.ExitError
 	if errors.As(err, &exit) {
-		return exit.ExitCode(), false, nil
+		// Windows gives exit codes as unsigned 32-bit numbers: -1, the code
+		// of a game ended from Task Manager, comes as 4294967295.
+		return int(int32(exit.ExitCode())), false, nil
 	}
 	return 0, false, err
 }
